REFACTOR : WIP Refactor Phase 1
