+++ install/DuckHunt.install4j
@@ -97,7 +97,7 @@
           <versionLine x="20" y="40" text="version ${compiler:sys.version}" font="Arial" fontSize="8" fontColor="0,0,0" fontWeight="500" />
         </text>
       </splashScreen>
-      <java mainClass="game.DuckHunt" vmParameters="${compiler:APP1_VM_PARMS}" arguments="" allowVMPassthroughParameters="true" preferredVM="" bundleRuntime="true">
+      <java mainClass="game.DuckHuntRunner" vmParameters="${compiler:APP1_VM_PARMS}" arguments="" allowVMPassthroughParameters="true" preferredVM="" bundleRuntime="true">
         <classPath>
           <scanDirectory location="lib" failOnError="true" />
           <directory location="etc" failOnError="true" />
