+++ src/game/DuckHuntRunner.java
@@ -1,0 +1,23 @@
+package game;
+
+import java.awt.*;
+import java.awt.event.*;
+
+import javax.swing.*;
+
+public class DuckHuntRunner {
+
+
+
+	public static void main(String[] args) {
+		// AudioClip au;
+		// URL codeBase = new URL
+		// ("http://vgmusic.com//music//console/nintendo/nes/Duck_Hunt.mid");
+		// au = getAudioClip(codeBase);
+		GameBuilder gameBuilder = new GameBuilder();
+
+	}
+
+}
+
+
